Getting tests working and getting close to 100% coverage
Increasing the amount of tests for the core app (flexible_content).
Removing all forms project-wide except for one for VideoForm that hits
YouTube and Vimeo APIs to verify a video_id. Added tests for that API's
reliability.

Also added code to auto-generate a form class if a type doesn't specify
one in its FlexibleContentInfo class or override get_form_class() to
return it. This is tested as well.

Updating settings and requirements to use coverage/django-coverage.

Updating utils to use simpler django import structure (app.ModelName)
and removing traces of old project name.